Factor predefined prompts into a separate file
